product modal to app
